Backplane BOM, FPGA RGB control error
Backplane BOM had incorrect parts for the BNC and Battery connector
Fixed spelling error in Eagle note
There were errors when the backplane controlled an RGB display via serial control. The serial circular buffer was not being filled correctly at wrap around. Code was changed so that the buffer was filled differently, one value per clock, with one increment of the pointer per clock. This seems to fix the problem.